Refactor: Move redis cache from utils to services
